requirement.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
requests
pydub
python-dotenv